refactor: remover testes e mocks obsoletos para casos de uso de usuário e login
