+++ back/src/__test__/src/utils/hashPassword.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect } from 'vitest';
-import { hashPassword } from '../../../../src/utils/password-generator'; // Ajuste o caminho conforme necessário
+import { hashPassword } from '@utils/password-generator'; 
 import bcrypt from 'bcrypt';
 
 describe('hashPassword', () => {
@@ -7,7 +7,6 @@ describe('hashPassword', () => {
     const senhaOriginal = 'minhaSenha123';
     const hash = await hashPassword(senhaOriginal);
 
-    // Verifica se é uma string e se o hash bate com a senha original
     expect(typeof hash).toBe('string');
     const corresponde = await bcrypt.compare(senhaOriginal, hash);
     expect(corresponde).toBe(true);
