+++ back/vitest.config.ts
@@ -1,4 +1,5 @@
 import { defineConfig } from 'vitest/config'
+import path from 'path'
 
 export default defineConfig({
   test: {
@@ -6,6 +7,16 @@ export default defineConfig({
     environment: 'node',
     coverage: {
       reporter: ['text', 'html'],
+    },
+  },
+  resolve: {
+    alias: {
+      '@utils': path.resolve(__dirname, 'src/utils'),
+      '@usecases': path.resolve(__dirname, 'src/usecases'),
+      '@domain': path.resolve(__dirname, 'src/domain'),
+      '@infrastructure': path.resolve(__dirname, 'src/infrastructure'),
+      '@main': path.resolve(__dirname, 'src/main'),
+      '@interfaces': path.resolve(__dirname, 'src/interfaces'),
     },
   },
 })
